+++ wxc_sdk/all_types.py
@@ -288,6 +288,8 @@ from wxc_sdk.telephony.vm_rules import BlockContiguousSequences, BlockPreviousPa
 from wxc_sdk.telephony.voice_messaging import MessageSummary, VoiceMailPartyInformation, VoiceMessageDetails
 from wxc_sdk.telephony.voicemail_groups import VoicemailGroup, VoicemailGroupDetail
 from wxc_sdk.telephony.voiceportal import ExpirePasscode, FailedAttempts, PasscodeRules, VoicePortalSettings
+from wxc_sdk.telephony.webhook_interest_registrations import Interest, InterestActor, InterestResource, \
+    WebhookInterestRegistration
 from wxc_sdk.tokens import Tokens
 from wxc_sdk.webhook import Webhook, WebhookCreate, WebhookEvent, WebhookEventData, WebhookEventType, \
     WebhookResource, WebhookStatus
@@ -392,11 +394,12 @@ __all__ = ['AIRDirectLineCallerIdName', 'AIRIntent', 'AIRSessionState', 'AIRSess
            'InProgressDevice', 'Incident', 'IncidentUpdate', 'IncomingPermissions', 'InitiateMoveNumberJobsBody',
            'InputMode', 'IntentTransferTo', 'InterceptAnnouncements', 'InterceptNumber', 'InterceptSetting',
            'InterceptSettingIncoming', 'InterceptSettingOutgoing', 'InterceptTypeIncoming', 'InterceptTypeOutgoing',
-           'InternalDialing', 'InterpreterForSimultaneousInterpretation', 'Invitee', 'InviteeForCreateMeeting',
-           'JobError', 'JobErrorItem', 'JobErrorMessage', 'JobExecutionStatus', 'JoinMeetingBody',
-           'JoinMeetingResponse', 'KemKey', 'KemModuleType', 'KnowledgeBase', 'KnowledgeBaseDocument',
-           'KnowledgeBaseDocumentDetails', 'KnowledgeBaseDocumentKnowledgeType', 'KnowledgeBaseDocumentStatus',
-           'KnowledgeBaseMappedBot', 'LargeOrgStatus', 'LayoutMode', 'License', 'LicenseProperties', 'LicenseRequest',
+           'Interest', 'InterestActor', 'InterestResource', 'InternalDialing',
+           'InterpreterForSimultaneousInterpretation', 'Invitee', 'InviteeForCreateMeeting', 'JobError',
+           'JobErrorItem', 'JobErrorMessage', 'JobExecutionStatus', 'JoinMeetingBody', 'JoinMeetingResponse',
+           'KemKey', 'KemModuleType', 'KnowledgeBase', 'KnowledgeBaseDocument', 'KnowledgeBaseDocumentDetails',
+           'KnowledgeBaseDocumentKnowledgeType', 'KnowledgeBaseDocumentStatus', 'KnowledgeBaseMappedBot',
+           'LargeOrgStatus', 'LayoutMode', 'License', 'LicenseProperties', 'LicenseRequest',
            'LicenseRequestOperation', 'LicenseUser', 'LicenseUserType', 'Lifecycle', 'LineKeyLabelSelection',
            'LineKeyLedPattern', 'LineKeyTemplate', 'LineKeyTemplateAdvisoryTypes', 'LineKeyType', 'LinkRelation',
            'Location', 'LocationAddress', 'LocationAndNumbers', 'LocationAssignedNumber', 'LocationCallCaptions',
@@ -504,11 +507,11 @@ __all__ = ['AIRDirectLineCallerIdName', 'AIRIntent', 'AIRSessionState', 'AIRSess
            'VoicemailGroup', 'VoicemailGroupDetail', 'VoicemailMessageStorage', 'VoicemailNotifications',
            'VoicemailSettings', 'VoicemailTransferToNumber', 'VolumeSettings', 'WaitMessageSetting', 'WaitMode',
            'WebexGroup', 'WebexGroupMeta', 'WebexGroupOwner', 'WebexStatus', 'WebexUser', 'WebexUserMeta', 'Webhook',
-           'WebhookCreate', 'WebhookEvent', 'WebhookEventData', 'WebhookEventType', 'WebhookResource',
-           'WebhookStatus', 'Week', 'WelcomeMessageSetting', 'WifiAuthenticationMethod', 'WifiCustomization',
-           'WifiNetwork', 'WorkSpaceType', 'Workspace', 'WorkspaceCalling', 'WorkspaceCallingHybridCalling',
-           'WorkspaceEmail', 'WorkspaceHealth', 'WorkspaceHealthIssue', 'WorkspaceHealthLevel',
-           'WorkspaceIndoorNavigation', 'WorkspaceLocation', 'WorkspaceLocationFloor', 'WorkspaceNumbers',
-           'WorkspacePersonalizationTaskResponse', 'WorkspaceSupportedDevices', 'WorkspaceWebexCalling',
-           'WrapUpReason', 'WrapUpReasonDetails', 'WrapupReasonQueue', '_Helper', 'dt_iso_str', 'enum_str', 'plus1',
-           'to_camel', 'webex_id_to_uuid']
+           'WebhookCreate', 'WebhookEvent', 'WebhookEventData', 'WebhookEventType', 'WebhookInterestRegistration',
+           'WebhookResource', 'WebhookStatus', 'Week', 'WelcomeMessageSetting', 'WifiAuthenticationMethod',
+           'WifiCustomization', 'WifiNetwork', 'WorkSpaceType', 'Workspace', 'WorkspaceCalling',
+           'WorkspaceCallingHybridCalling', 'WorkspaceEmail', 'WorkspaceHealth', 'WorkspaceHealthIssue',
+           'WorkspaceHealthLevel', 'WorkspaceIndoorNavigation', 'WorkspaceLocation', 'WorkspaceLocationFloor',
+           'WorkspaceNumbers', 'WorkspacePersonalizationTaskResponse', 'WorkspaceSupportedDevices',
+           'WorkspaceWebexCalling', 'WrapUpReason', 'WrapUpReasonDetails', 'WrapupReasonQueue', '_Helper',
+           'dt_iso_str', 'enum_str', 'plus1', 'to_camel', 'webex_id_to_uuid']